int to unsigned

+++ LockFree/Source/LFIntQueue.cpp
@@ -4,6 +4,7 @@
 #include <atomic>
 #include "prnPrint.h"
 
+#include "LFIndex.h"
 #include "LFIntQueue.h"
 
 using namespace std;
@@ -17,9 +18,9 @@ namespace LFIntQueue
 
    typedef struct
    {
-      int         mValue;
-      atomic<int> mQueueNext;
-      atomic<int> mListNext;
+      int                  mValue;
+      atomic<unsigned>     mQueueNext;
+      atomic<unsigned>     mListNext;
    } QueueListNode;
 
    static const int cInvalid = 0x8000000;
@@ -29,19 +30,19 @@ namespace LFIntQueue
    //***************************************************************************
    // Queue Members
 
-   atomic<int> mQueueHead;
-   atomic<int> mQueueTail;
+   atomic<unsigned> mQueueHead;
+   atomic<unsigned> mQueueTail;
 
    //***************************************************************************
    //***************************************************************************
    //***************************************************************************
    // Free List Members
 
-   bool listPush(int  aIndex);
-   bool listPop(int* aIndex);
+   bool listPush(unsigned  aIndex);
+   bool listPop(unsigned*  aIndex);
 
-   atomic<int> mListHead;
-   atomic<int> mListSize;
+   atomic<unsigned> mListHead;
+   atomic<unsigned> mListSize;
 
    //***************************************************************************
    //***************************************************************************
@@ -52,9 +53,9 @@ namespace LFIntQueue
    static QueueListNode* mNode = 0;
 
    // Number of blocks allocated
-   static int mAllocate = 0;
-   static int mQueueAllocate = 0;
-   static int mListAllocate = 0;
+   static unsigned mAllocate = 0;
+   static unsigned mQueueAllocate = 0;
+   static unsigned mListAllocate = 0;
 
    //***************************************************************************
    //***************************************************************************
@@ -73,14 +74,14 @@ namespace LFIntQueue
 
    void initialize(int aAllocate)
    {
-      mAllocate      = aAllocate;
+      mAllocate      = (unsigned)aAllocate;
       mQueueAllocate = aAllocate + 1;
       mListAllocate  = aAllocate + 1;
 
       if (mNode) free(mNode);
       mNode = new QueueListNode[mListAllocate];
 
-      for (int i = 0; i < mListAllocate - 1; i++)
+      for (unsigned i = 0; i < mListAllocate - 1; i++)
       {
          mNode[i].mValue = 0;
          mNode[i].mQueueNext = cInvalid;
@@ -95,7 +96,7 @@ namespace LFIntQueue
       mListSize = mListAllocate;
       mListHead = 0;
 
-      listPop((int*)&mQueueHead);
+      listPop((unsigned*)&mQueueHead);
       mQueueTail = mQueueHead.load();
 
       mWriteRetry = 0;
@@ -146,7 +147,7 @@ namespace LFIntQueue
    {
       // Try to allocate a node from the free list.
       // Exit if it is empty.
-      int tNode;
+      unsigned tNode;
       if (!listPop(&tNode)) return false;
 
       // Initialize the node with the value.
@@ -154,7 +155,7 @@ namespace LFIntQueue
       mNode[tNode].mQueueNext = cInvalid;
 
       // Attach the node to the queue tail.
-      int tTail,tNext;
+      unsigned tTail,tNext;
       mWriteRetry--;
       while (true)
       {
@@ -194,7 +195,7 @@ namespace LFIntQueue
    bool tryRead(int* aValue)
    {
       // Store the head node in a temp.
-      int tHead, tTail, tNext;
+      unsigned tHead, tTail, tNext;
       mReadRetry--;
       while (true)
       {
@@ -230,13 +231,13 @@ namespace LFIntQueue
    // Insert a node into the list before the list tail node.
    // There can be no concurrent calls to this.
 
-   bool listPush(int aNode)
+   bool listPush(unsigned aNode)
    {
       // Exit if the list is full.
       if (mListSize >= mAllocate) return false;
 
       // Store the head node in a temp.
-      int tHead = mListHead;
+      unsigned tHead = mListHead;
       mPushRetry--;
       while (true)
       {
@@ -259,11 +260,11 @@ namespace LFIntQueue
    // This detaches the node that is before the tail list node.
    // There can be concurrent calls to this.
 
-   bool listPop(int* aNode)
+   bool listPop(unsigned* aNode)
    {
       // Store the head node in a temp.
       // This is the node that will be detached.
-      int tHead = mListHead;
+      unsigned tHead = mListHead;
       mPopRetry--;
       while (true)
       {
